requirements.txt
-e .
Django==1.8.8
django-tagging==0.4
django-voting==0.2
dj-database-url==0.3.0
factory-boy==2.6.0
psycopg2==2.6.1
